requirements.txt
torch
torchvision
torcheval
scikit-learn
pillow
streamlit
matplotlib
numpy
pyngrok
seaborn
tqdm